move to a more coherent use of standard integer types
try to use
- size_t for everything that might end up as an index
- {u}intptr_t for integers that are to be compatible with pointers
- {u}intmax_t when we don't know how large things might get

+++ include/orwl_enum.h
@@ -42,7 +42,7 @@
 #define DECLARE_ENUM(T, ...)                                            \
 typedef enum { __VA_ARGS__ ,                                            \
                T ## _amount,                                            \
-               T ## _max = ((unsigned long)(T ## _amount) - 1ul),       \
+               T ## _max = ((size_t)(T ## _amount) - 1u),               \
                T ## _min = 0                                            \
 } T;                                                                    \
 extern char const* _ ## T ## _names[T ## _amount];                      \
